+++ api/api/puzzle_batch_edit.py
@@ -157,4 +157,4 @@ class AdminPuzzleBatchEditView(MethodView):
         # TODO: if action in ('reject', 'delete'): #Also apply to any puzzle instances
 
         cur.close()
-        return make_response("204", 204)
+        return redirect("/chill/site/admin/puzzle/", code=303)